LG-10124: refactor based on jira card update.
