+++ lib/telephony/alert_sender.rb
@@ -1,5 +1,5 @@
 module Telephony
-  class AlertSender
+  class AlertSender < BasicSender
     SMS_MAX_LENGTH = 160
 
     def send_account_reset_notice(to:, country_code:)
@@ -44,38 +44,6 @@ module Telephony
       response = adapter.send(message: message, to: to, country_code: country_code)
       log_response(response, context: __method__.to_s.gsub(/^send_/, ''))
       response
-    end
-
-    private
-
-    def adapter
-      case Telephony.config.adapter
-      when :pinpoint
-        Pinpoint::SmsSender.new
-      when :test
-        Test::SmsSender.new
-      end
-    end
-
-    def log_response(response, context:)
-      extra = {
-        adapter: Telephony.config.adapter,
-        channel: :sms,
-        context: context,
-      }
-      output = response.to_h.merge(extra).to_json
-      Telephony.config.logger.info(output)
-    end
-
-    def log_warning(alert, context:)
-      Telephony.config.logger.warn(
-        {
-          alert: alert,
-          adapter: Telephony.config.adapter,
-          channel: :sms,
-          context: context,
-        }.to_json,
-      )
     end
   end
 end
